Try to make function chain a big more logical
Previously:
fetchSetAndMaterials has the page number and the page size. From
those it builds a url and calls
fetchMaterialsFromSetByUrl, which would then extract the details
from the url because it needed them. Better to issue the request
from fetchSetAndMaterials, where the details are known.

Also changed the regex for extracting page number from url
use \d+ instead of \d*, because if the page number is missing,
it won't work.