Bugfix - prevent set shaper auto-requesting set contents without page number
It seems like when a set is partway through loading, getSelectedTop/BottomPage might not return anything
useful, and that ends up with a set request being sent without a page number.
Instead, if there is no page number, do not try and reload. A load is presumably already in progress.